Add a token to all URLs that generate and return files
Till now all the URLs that generate files (workout PDF, etc.) checked
that the user is the owner of the object. This does break down when the
application is "installed" and opened as an app. While there is code in
the android app for this, this is a more general solution that works
acrross all devices (firefox OS, desktop, etc.). These tokens allow
'anonymous' access to the URLs and while not really one-time tokens,
they offer enough security for this.

Fixes #80