Create guest users on dashboard or demo entries
This change makes the WgerAuthenticationMiddleware more constrained, as it only
creates guest users when arriving directly at the dashboard and not on basically
every other URL that had user data. This is done in preparation of the read-only
access to logs or workouts, as in this case no guest user should be created.