Don't allow adding images during group creation
When creating a group, there is no PK at the time the image is saved so
the image gets saved under the "group-images/None/<uuid>" path. Editing
makes no problem. Because of this, the create dialog can be opened again
in a modal dialog and the form enctype doesn't have to be set.