Return muscle images from model
Each muscle is associated with a primary and secondary image, which are
used if it is the primary or secondary muscle in an exercise,
respectively.

The previous code computed the URLs of these images in the
render_muscle() function. This commit generalizes this slightly by
moving that calculation to the muscle model instead.

Doing so cleans up the code slightly, and also makes it easier to
return the image URLs as part of the wger API.