Add a global las update date for exercises
This is simply the last date for the exercise, any of the translations, videos
or images. Also, we're renaming the date entries for some of the models. Things
like the workout routines, etc. still need to be changed, but those would break
the apps and need a bit more of planning